Next/Previous buttons on half the screen
Lots of fun learning ConstraintLayout: it is like RelativeLayout but with hints
of springs and struts, and other mobile platforms.

Got two buttons nex/previous on half the screen that are shown but with zero
alpha.  They allow advancing the image.

Tapping on them also shows the System UI, because System UI is usually
at the borders and shouldn't disturb with the images.

Now next step should be filesystem traversal and reading images from /sdcard.