File reading!
Reading from disk, displaying actual images on the view.

Most fun ever.

Portrait images are displayed incorrectly because the orientation on the file
is not being read.  The navigation bar has been removed entirely because it
wasn't adding anything.  Need some way to show a bar overlaid on the screen.