Rotation works but still OOMs
The byte[] object is about 100Mb, which means that I'm not
sampling the bitmap, just reading all of it directly. This
also means that any time two bitmaps are concurrently loaded,
the application crashes since the previous bitmap is needed
to back the imageview, and the graphics memory uses about 100M
as well.

I need to subsample the image before loading it.