Fixed memory crash and portrait centering
It worked to think about the problem in isolation.
I changed the Bitmap loading code in a single edit,
subsampling the Bitmap and then rotating it on the
pivot point being the middle of the image. Then this
is assigned to the ImageView with FIT_CENTER, and
this fixes the offset for the portrait images, and
also fixed the memory crashes when iterating quickly
through them.