Removing superfluous code
Now that the code with in-place rotation and scaling is working, I am
removing the earlier, buggy code.

Memory measurements show average memory usage on a Pixel C as 170Mb,
with 50Mb java, 25Mb native, 75Mb graphics memory usage.

A Pixel C has a huge resolution: 2560x1800 for the imageView, since
there is no surrounding space. This increases the memory usage
considerably. In addition, I have very high quality photos from a DSLR
and so they have to be downsampled to be shown on this screen.

Need to be tested on other tablets as well, especially lower-end
tablets.