Intermediate code: ZIP file downloads
Added code to download a zip file and then to unzip it into the
individual files. There are lots of half-baked things, poor variable
names and other ugliness in there. At a minimum, I have confirmed that
the code compiles and downloads the zip file, but the unzipping is not
tested yet.