Unzipping really works
In the previous commit, I was incorrectly writing out the comment
field rather than the zipfile data. Now I am correctly reading the
bytes associated with each file, and writing it out, a 4Mb chunk at a
time. Verified with the files zipped with Linux earlier that it does
extract all three files with the correct size.

$ ls -l /tmp/2018_Feb_02/
total 10024
-rwxr-xr-x  1 user  wheel  1881696 Nov 27 19:12 _DSC6603.JPG
-rwxr-xr-x  1 user  wheel  1571765 Nov 27 19:12 _DSC6609.JPG
-rwxr-xr-x  1 user  wheel  1674574 Nov 27 19:12 _DSC6635.JPG

$ adb shell ls -l /sdcard/Pictures/test
total 10024
-rw-rw---- 1 root sdcard_rw 1881696 2018-12-01 20:58 _DSC6603.JPG
-rw-rw---- 1 root sdcard_rw 1571765 2018-12-01 20:58 _DSC6609.JPG
-rw-rw---- 1 root sdcard_rw 1674574 2018-12-01 20:58 _DSC6635.JPG

There are still a lot of issues: the directory is called 'test' which
is not proper. Ideally the directory name should be read from a
manifest file, or some metadata.

Another problem is that I'm still reading the previously downloaded
file, since DownloadManager returns too early for me to be able to
read the current file.