A wait is required
Either reading the ContentProvider URI or the file results in zero
byte files initially. Either I am doing too much on the main thread in
the CPU, or there is just a long lag in a 4Mb file showing up on the
filesystem.

The solution is to check periodically till the file has a nonzero
size. Can do this tomorrow.