Wait for download to settle
This patch waits for the file download to settle down by waiting a
second at a time till the ParcelFileDescriptor has nonzero size.

Ideally I should wait for the size to be compared against what we
expect, not just to be non-empty.

I should also add checksum verification and decryption which should be
required, and not optional.