Added a custom scheme
The custom scheme photoviewer://eggwall/anything?q=stuff now works
with this app. This has the advantage that I can now author web pages
that takes the user to the photo viewer application, and allows them
to download the zip file directly.

Here's some sample html code that works today:
This is <a href="photoviewer://eggwall/test?q=this">a test </a>.

Done correctly, this allows a single link that is a feed that the
program can routinely monitor for changes. When new content is
available, it can download the zip files and alert on the device.

I'll add the feed tomorrow. I'll also try having the full end-to-end
file download, decryption, unzipping, and photo viewing on it, because
all the pieces are in place.

Another minor revision is that the Slideshow icon changes to a pause
button when the slideshow is running and a play button when paused.