Added a MainController
The controller objects are getting unwieldy and large. Something needs
to orchestrate their functionality and reduce each controller knowing
about the other.

This is going to be the job of the MainController, to orchestrate the
FileController, the NetworkController and the UiController, hopefully
keeping their implementations independent of the other objects.