UiController is unaware of File management now
This has a much cleaner structure, as the UiController just handles
button presses and gestures, and asks the MainController to take
actions.

Behind the scenes, the MainController calls the UiController to handle
next images, but it does so by querying the FileController to get the
correct next/previous file.

The goal is to make every controller dependent only on the
MainController and the leaf controllers are unaware of each others'
presence.