Album database hook-up
Album numbers are now created in databases, and each download goes
to a unique directory.

Encrypted zips should work, though I haven't tested them out.
Zip files are deleted after extraction.

The entire FileController.createUnzipper() needs to be called from a
background thread because it needs a unique album.id and it can only
get that on an insert. This has the subtle bug that you could call
MainController.download() in quick succession and get two album
inserts in the same location. Let's assume that all album creation is
done by the user and so it is very unlikely to happen. This will be a
problem if download() is being called programmatically.

Testing with encrypted archives should be next.