Downloading to unique location works
Downloads are to unique locations rather than test.zip and they all
unzip correctly.

Switching to the newly downloaded gallery does not work yet. I suspect
the FileController.setDirectory code is pretty ancient and didn't work
well. It would give random troubles when starting afresh, so this is a
good opportunity to read the database instead of the filesystem.

Decryption has not been tried out yet.