Allow http requests (as opposed to httpS only)
Tested it on the latest Android version, and realized that I was
hitting 'cleartext not permitted' errors.

Turns out that cleartext support is disabled by default for
DownloadManager. Since I am encrypting the entire file, cleartext
transport should be fine.

https://developer.android.com/training/articles/security-config#CleartextTrafficPermitted

So I fixed that.

I still ran into issues where the directory was not created when the
program was just installed, and I need to get to the bottom of why the
/sdcard/Pictures/eggwall directory was not created.

After enabling cleartext support, the device correctly downloaded the
encrypted .asc file and decrypted it (because the password is in
cleartext in the source code).

Adding album selection support, better iconography, and Key import
very soon.