Created icons
I created an icon in Photoshop and exported it using Roman's website:
https://romannurik.github.io

The icon is a simple bulls-eye. Newer versions of Android force a
circular icon, which I really dislike. But given that we are going to
use circular icons, we might as well get the most catchy, most
appealing circular icon.

I can modify it later, I have the PSD file for later modifications.