Allow Secret Keys to be imported with /import?key=...
I've split the URL scheme so that downloads happen with
/download?src=.. and key imports happen with /import?key=..

This allows for keys to be read from URLs, so the sharer starts out
sending a secret key, tagged with a uuid for that key.
Then, they send packages encrypted with that key and URLs generated
with that package mark the uuid for the key that the package
corresponds to.

It hasn't been tested extensively, but it compiles, there is a
database table for the keys, and I have the capability to insert
keys into the database.

Keys are not currently read, there is a single key still hardcoded
into FileController.Unzipper.handleFile().

But this is good enough to be checked in and bugs get fixed in
subsequent check-ins.