Decryption with keys from database!
Added functionality to insert keys into the database with /import and
the functionality to pick up the correct key with a uuid during
/download

Tested this out on files hosted on local machines.

Also added a routine to print debug messages which greatly speeds up
progress as I don't have to read logcat and can instead see Toasts
instead when something succeeds or fails.

Next up:

1. To show all the albums by looking through the database rather than
   reading disk, even on app startup.

2. More cleanup on state display to users, so they know that downloads
   succeeded or failed.

3. A deeper look into which methods can be called on the UI thread,
   and which need to be called off the UI thread.

4. An initial screen, which many users will see.