Background threads done right
MainActivity.onCreate() called file handling in the past. Now we
solemly create a background thread for doing that.

updateImage() was a mess of UI and background thread calls, so instead
of modifying all callsites, I changed the method to accept calls from
any thread, and dutifully pop into a background thread if called from
the UI thread, and stay in its place if called from a background thread
(to avoid creating too many threads recklessly)