Fix some background thread/foreground thread issues
Also, create a method to crashHard(String message) so that we can
truly bail in exceptional circumstances.

I need to find these occurrances and see what the best alternative
to crashing can be. In some cases, like creation of a MainController,
there might be no alternative, and crashing is actually a feature.