Read the database for the initial album
All this while I was reading the disk in a very not-cool way.  We knew
how the files are stored, and so let's just check the location and
point to that location.

This commit looks for an album on the database and points to the most
recently viewed album.

Some other clean-up has gone into it, but the biggest part has been
the MainController.showInitial() work.