Resumes state when restarted
On rotation, or when the application is killed and restarted, it tries
(its best) to resume the state it had last. It tries to load the same
album ID it had earlier, and the same offset.

This largely works.

Moved some code about, added a lot of comments, and also cleaned a lot
of old code.

The save/resume has now brought back the OOM errors. Should debug those
next.