Fixed the memory bug on rotation
The bug was caused by UiController.sampling() method returning very
large values when the view was *just* instantiated. I suspect this is
because the layout has not been done and so the bounds are not available.

As a result, unscaled, full-size bitmaps were created, and not GC'd on
immediate rotation.

Here's a trace that displays the problem. See how the sample size
rolls over and doesn't get anything decent:

10:13:25.922 D/UiController: updateImage sample size = -2147483648
10:13:25.923 D/UiController: updateImage called with free memory(bytes) = 951091200
10:13:41.116 D/UiController: updateImage sample size = -2147483648
10:13:41.117 D/UiController: updateImage called with free memory(bytes) = 723476480
10:13:46.532 D/UiController: updateImage sample size = 2
10:13:46.535 D/UiController: updateImage called with free memory(bytes) = 726634496
10:13:47.596 D/UiController: updateImage sample size = 2
10:13:47.597 D/UiController: updateImage called with free memory(bytes) = 769789952
10:13:48.601 D/UiController: updateImage sample size = 2
10:13:48.602 D/UiController: updateImage called with free memory(bytes) = 866471936
10:13:49.378 D/UiController: updateImage sample size = 2
10:13:49.379 D/UiController: updateImage called with free memory(bytes) = 868007936
10:13:51.471 D/UiController: updateImage sample size = -2147483648
10:13:51.472 D/UiController: updateImage called with free memory(bytes) = 846094336

The temporary fix is to clamp the sampling size to something
reasonable (4 for now) if the sampling method returns bad values.

The real fix would be to check the screen bounds in some other way so
that the correct sampling() value is returned.