Making the scaling code more resilient
When the Activity is just created, the imageView has (0,0) bounds.
This causes a problem for the UiController.scaling() method since it
cannot fit an image in a zero-sized space.

One alternative is to use the screen geometry as reported by
DisplayMetrics as an estimate. The scaling code can still report a
number to scale the image by.

This allows us to remove the 400ms delay in Activity start, and
eliminates the crash altogther. More testing is needed to confirm that
this is correct under all Activity transitions.

I have also started adding code for selecting an Album, but I might
need to create a different Activity instead.

Slideshow viewing seems to have regressed.

And rotation preserves state, but starting the Activity from the
Launcher doesn't preserve state.

It still crashes, but relatively rarely on a combination of slideshow
and rotations.