Adding a new Activity for showing the Album list
Currently, the Activity just shows a nasty, ugly list, but with time
I can add more to it (last download date, some sample images, time of
download, etc)

No selection is allowed, and it also doesn't show an actual list.

On a tablet the Activity looks hideously ugly, but with a proper
layout file, and a transparency along one wall, it might just blend in
with the view behind it.

To be really fancy, I could make Fragments that display the list
properly, or even add the RecyclerView to the vertical 'start' of the
display, and push out the image view.

For now, this works great for prototyping and ensures that no code
mingling between the two occurs.

Also fixed the bug with slideshow functionality. Slideshows work
again.