Cleanup of UI
Discovered a bug with RTL layouts. Our icons are left/right pointing
arrows, and the direction that images advance is pretty arbitrary
anyway. So made the right pointing arrow point always be on the right
end of the screen. And left pointing arrow always on the left end of
the screen.

Added some comments to remember what this UI is doing.

And threw out some DrawerListener code that is of dubious utility.