Bitmaps (and UiControllers) being retained.
On five rotations, I see that older UiController objects are still
retained. This means that image objects (and their corresponding
byte[] bitmap objects) are still retained.

This usually points to some Activity context being preserved, which
then holds a reference to the MainController, which has a reference
to the UiController, which has a bitmap.

Each bitmap itself is 24M, so I would improve the system significantly
if I find what's holding on to the Context and eliminate that.

The Java memory is being held, and also the graphics memory suggesting
that the old Activity object is being retained as well.