Null out the objects on destroy
On an Android tablet, six copies of the Activity are kept
around. Calling gc (through the Android Studio UI) the first time gets
rid of some of them and calling gc again leaves just one Activity.

So it is clear that somewhere either memory references are retained
for optimization, or that there is dangling reference.

I'm nulling out the MainController and its child controllers'
references onDestroy() to avoid keeping object cycles in memory, in
case that is making the problem worse.

On my tablet, this fix does not reduce memory consumption, but it also
doesn't make anything bad.

On a phone, running Android P the entire app is very crashy, so that
is a good device to do some field testing and methodically remove all
memory bugs and issues.