Bounded memory usage! Slideshow works!
Instead of a single Bitmap, I allocate two and craftily switch between
them so one is displayed while the other is used to load the next
image.

This takes twice the ram (2x24MB of RAM on my tablet)

The advantage of this method is that it only takes 2x24MB and memory
usage is totally bounded. With the previous change, rotation and image
loading will never run out of memory. If you can load two images, one
after another, you are done with the allocations for the life of the
program.

Since the byte[] objects are so long and contiguous, this is a huge
win.

Also fixed slideshow, so it resumes from the previous Bundle and
resumes correctly.

This version is considerably more stable on my phone as well, and
doesn't crash.

Initial directory creation and Album import is still buggy, and I need
to add a gallery list picker.