Reduce code visibility, improve naming
Also adding a call to dismissIntroScreen() on updateImage, which is
when the intro-screen (if any) should be removed.

I haven't tested all codepaths to see if the fling detection and other
handlers are correctly restored, and there might be UI state I have
forgotten to restore (GONE -> VISIBLE missing), or touch handlers that
aren't correctly instantiated again.

Leaving this as it is because it is good enough.

Now time to add other functionality.