Fixed: Dark screen on phones
On phones (tall screens / wide screens), where the aspect ratio is
more rectangular than square, we run into an interesting artifact:

Let's say the screen is in landscape mode. A portait image will get
sampled aggressively (say 1/4 the original length and 1/4 the original
width). And along the sides you get letterboxing: black areas of no
content.

In that same byte[] buffer, if we were to unpack a landscape screen at
lower sampling (say 1/2 the original length and 1/2 the original
width) then it won't fit into that buffer, since the image is
considerably larger.

BitmapFactory.decodeFile(...) in this situation helpfully returns a
null value, which shows a blank screen. Better than crashing, but not
very helpful.

This change tries two approaches:

1. Create a new buffer, ignoring a previously allocated buffer.
2. Sample the image more, throwing out more information reducing the size.

If this is perfectly programmed, #1 should only happen twice (since
there are two buffers we switch between). This is not the observed
behavior, and the allocation happens repeatedly. But at least the null
return value is gone, and so the behavior is better.