Clean up the background code in BroadcastReceiver.onReceive()
All that background code really belongs in a separate method. Also,
now that the error code is in one place (and in the background
thread), it can be cleaned up, and the AsyncTask that was used earlier
can be eliminated.

This finishes the cleanup of BroadcastReceiver.onReceive() in
NetworkController. The code is good enough, guarded well enough and
can catch most errors reasonably gracefully. Some of these errors can
no doubt be turned into AndroidRoutines.crashDuringDev() but these
errors don't happen frequently.