Removed the toolbar.
Experimented with the behavior if the toolbar and associated runnables
were all gone, and the performance is just as good on the two devices
I tested. They were hidden elements, and we weren't showing anything
on them.

This might need to be brought back, so this is an isolated change by
itself.

Needs more testing on other devices and Android versions.