Show a list of all albums in the Drawer!
This change lists out all the albums available in the database in the
drawer. It doesn't yet allow switching to the album, and it also
doesn't display when the album was last viewed.

It also highlights how the viewer should not download the same
location twice because I have six instances of the same album right
now.

But the skeleton is all in place.