Selection of albums works
An album list is shown in the drawer, and selecting on the list
displays that album. The UI is a bit clunky right now but it works,
and accurately displays the content.

What is needed is more metadata for each menu item so the user can
see last-viewed, or other information about the album.