Tested with old tablet running L
And it works great!

Cranking down the minimum SDK version to 17, and it still compiles
fine.  I might not have any devices that old, but it is good to know
that it runs fine on six year old hardware just fine.

Imported the secret key, and imported the existing album as well.