Update the drawer on the main thread
With this, refreshAlbumList() is complete. The method removes the
previous albums from the drawer, it refreshes the list, and modifies
the UI in the main thread.

I hit an interesting bug on Fire tablets where mDrawer.post(Runnable)
doesn't actually run when called in onCreate(). On Android devices they
run fine, so I'm going to leave this correct behavior around.

Another option is to call refreshAlbumList() a second or two after
onCreate() is called.

Tested on Chromebooks with Play Store as well, and it runs fine but
there is an intersesting issue there that the custom URL scheme
doesn't work as well as it does on phone. Instead of using custom URLs
exclusively, I should add support for Import'ing a string, which pulls
apart the location, the key name, etc. This should be easy to
do. Chromebooks were also incredibly slower than I expected.