Add a gallery import Activity
This activity will be used on devices that can't handle custom URLs,
and also for introducing what the program does.

This is a good place to include demonstration keys and demonstration
photo galleries as well.

Some clean-up in the Manifest as well because the move to androidx
messed that up.