ImportActivity downloads an album
The Activity allows for text input, passes it back to the caller, and
then the caller (MainActivity) downloads the URI, unpacking it and
including this in the database, as expected.

This should make the application usable on devices with no support for
custom URI strings, or for sending custom URIs by email that can be
copy-pasted back into the Aim app.

Some minor clean-up of the ImportActivity now that most of it works
correctly.

What's next:

1. I'd love to add more information to the ImportActivity that tells
   the user what the application does, and also some sample keys and
   albums they can download.

2. Make downloads unique. If the file-size is identical, and the key
   and album are identical, we should refuse to download that album
   again.

3. A URI that can be probed regularly for downloads, keys, etc.

4. Making this ImportActivity handle many different downloads, and key
   imports as well.