Handle key import and debugging commands in the Import Activity
Allowed more actions to be done in the Import Activity and cleaned out
the code interface quite a bit.

I also tested this on a Chrombook device with Play Store using adb
over tcpip.

On a Chromebook, using http doesn't work at all, and I have to rely on
https. For this purpose I have uploaded an encrypted image archive on
Dropbox and used that (with ?dl=1 to automatically download the image).

Works surprisingly well, and now I can switch to using a Chromebook
device for most of my testing.

At this point, this code works on ancient tablets, phones, and also on
current-gen phones and Chromebooks with Play Store enabled, all from
the same apk image. This is quite amazing.