Check for duplicates before downloading
Refuse to download a duplicate album, based on the URL and the name
provided.  These are all exact matches, so it is a light check that
the album was previously requested.

I should make this more rugged by restarting a download if it did not
succeed, and remove the previous entry.

Some more comments in the magic xml files.