Add support to monitor a URL
Once these tablets are showing images all day, it would be great for
them to automatically update their images.

The /monitor?beacon=url is a mechanism to do that.

This change adds support to remember a beacon. A subsequent change is
needed to act upon the beacon by polling it, downloading the content,
and then acting upon the content: which could be a key or a download.

Once that is working, I need to monitor for overall file size and then
prune the directory to keep filesystem consumption down.