Separate SharedPreferences into a separate class
The access to SharedPreferences all over the code was a bad idea, and
would lead to issues in the future. Made all access go through a
single entry-point to keep all access the same.

To modify a SharedPreference, the caller mentions which key is
modified, and what it is set to. No checking is done to ensure that
String values are written where Strings are expected.

This change keeps the defaults spread all over code, but there is no
clean way to keep int, String and other default values together in a
single structure, so I'm leaving it as-is for now.