Cleaned up the defaults for Settings
The defaults for the settings are stored in the Pref class, with a way
to set both integer and string defaults in place. This simplifies the
implementation of the caller, as they don't need to know what the
values should be.

Also made the code bullet-proof against ClassCastException which can
be thrown when the incorrect type is read.