Call the timer() frequently and purge disk
Purges disk when the timer is called every hour.

Calculates disk usage and keeps it below 100Mb. This setting needs to
be exposed through the SettingsActivity.

Allowed debugging timer functionality through a extra drawer button
called 'Refresh', but I will remove that at launch once the
functionality is working.

Need to add the periodic download, but it should be relatively easy.

I should also add the ability to delete an album directly, I could add
that as a separate option.