Got the download to work
Fixed a couple of issues to get the download on the beacon to work,
but now it cannot get the key UUID in the database.

All this points to the need for a better, more reliable PhotoEncrypter
tool that can take a zip file, upload it to dropbox, get a URI and
then create the correct URL because the entire process right now is
pretty error prone.

The beacon is plugged in, it downloads the content. And if an
unencrypted zip file were given, it should work.

More testing required tomorrow.