Call timer() on background workers for File and Network
File controllers and Network controllers will not be able to do
anything productive on the main thread. The UI controller will only be
productive on the main thread.

So FileController.timer() and NetworkController.timer() is called on a
WorkerThread and the the UIController.timer() is called on the main
thread.

Other minor cleanup.

In the testing on Nakasi, one thing popped out, which was how messy it
is to have a another activity handle the intent. That Activity should
close itself because the underlying activity is still around.
Let's implement that next.