Try: Keep a single Activity around.
Haven't had a chance to try this out, but checking it in.

Keep a single Activity resorting to launchMode and taskAffinity voodoo
to fall back to one Activity if it exists.

This makes a call to Activity.onNewIntent() when the Activity exists,
so implementing that and refactoring the previous call in onCreate()
so both methods call the same intent handling code.