Adding a TODO
Discovered that if the application is closed while the download is
still in progress then we will never recover from it:
1. We won't unpack the download, since the Activity was stopped.
2. We won't dowload that file again, since it was downloaded once.

The answer is to remember which files were requested, and check on
start-up if the files have been downloaded and then handle the
downloaded file.

But I'll park this for now, and work on the mechanism to create zip
files nicely.